perf: 优化插件代码结构。 fix: 修复卸载插件之后，线程无限自旋的问题。
